requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
python-dateutil==2.8.2
matplotlib==3.7.2
scikit-learn==1.3.0
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
python-dateutil==2.8.2
matplotlib==3.7.2
scikit-learn==1.3.0
pandas==2.0.3
flask-cors==4.0.0
python-speech-recognition==3.10.0
gtts==2.3.2
translate==3.6.1
python-speech-recognition==3.10.0
gtts==2.3.2
translate==3.6.1
